+++ src/dom/v7400/object/property.rs
@@ -1,14 +1,14 @@
 //! Object properties and related stuff.
 
 pub use self::{
+    loader::LoadProperty,
     node::{PropertyHandle, PropertyNodeId},
     object_props::ObjectProperties,
     properties::{PropertiesHandle, PropertiesNodeId},
-    value::LoadPropertyValue,
 };
 
+mod loader;
 pub mod loaders;
 mod node;
 mod object_props;
 mod properties;
-mod value;