+++ src/dom/v7400/object/property/loader.rs
@@ -1,9 +1,9 @@
-//! Property value.
+//! Property loader.
 
 use crate::dom::v7400::object::property::PropertyHandle;
 
-/// A trait for property value loader types.
-pub trait LoadPropertyValue<'a>: Sized {
+/// A trait for property loader types.
+pub trait LoadProperty<'a>: Sized {
     /// Value type.
     type Value;
     /// Error type.
@@ -16,12 +16,12 @@ pub trait LoadPropertyValue<'a>: Sized {
     fn load(self, node: &PropertyHandle<'a>) -> Result<Self::Value, Self::Error>;
 }
 
-impl<'a, T> LoadPropertyValue<'a> for &'_ T
+impl<'a, T> LoadProperty<'a> for &'_ T
 where
-    T: Copy + LoadPropertyValue<'a>,
+    T: Copy + LoadProperty<'a>,
 {
-    type Value = <T as LoadPropertyValue<'a>>::Value;
-    type Error = <T as LoadPropertyValue<'a>>::Error;
+    type Value = <T as LoadProperty<'a>>::Value;
+    type Error = <T as LoadProperty<'a>>::Error;
 
     fn expecting(&self) -> String {
         (*self).expecting()