+++ src/low/v7400/attribute/type_.rs
@@ -1,4 +1,4 @@
-//! Attribute type.
+//! Node attribute type.
 
 use std::io;