+++ src/pull_parser/v7400/attribute/loader.rs
@@ -1,59 +1,49 @@
-//! Node attribute visitors.
+//! Node attribute loader.
 
 use std::{fmt, io};
 
 use crate::pull_parser::{error::DataError, Result};
 
-pub use self::{
-    direct::DirectVisitor,
-    single::{ArrayVisitor, BinaryVisitor, PrimitiveVisitor, StringVisitor},
-    type_::TypeVisitor,
-};
-
-mod direct;
-mod single;
-mod type_;
-
-/// A trait for attribute visitor types.
-pub trait VisitAttribute: Sized + fmt::Debug {
+/// A trait for attribute loader types.
+pub trait LoadAttribute: Sized + fmt::Debug {
     /// Result type on successful read.
     type Output;
 
     /// Describes the expecting value.
     fn expecting(&self) -> String;
 
-    /// Visit boolean value.
-    fn visit_bool(self, _: bool) -> Result<Self::Output> {
+    /// Loads boolean value.
+    fn load_bool(self, _: bool) -> Result<Self::Output> {
         Err(DataError::UnexpectedAttribute(self.expecting(), "boolean".into()).into())
     }
 
-    /// Visit `i16` value.
-    fn visit_i16(self, _: i16) -> Result<Self::Output> {
+    /// Loads `i16` value.
+    fn load_i16(self, _: i16) -> Result<Self::Output> {
         Err(DataError::UnexpectedAttribute(self.expecting(), "i16".into()).into())
     }
 
-    /// Visit `i32` value.
-    fn visit_i32(self, _: i32) -> Result<Self::Output> {
+    /// Loads `i32` value.
+    fn load_i32(self, _: i32) -> Result<Self::Output> {
         Err(DataError::UnexpectedAttribute(self.expecting(), "i32".into()).into())
     }
 
-    /// Visit `i64` value.
-    fn visit_i64(self, _: i64) -> Result<Self::Output> {
+    /// Loads `i64` value.
+    fn load_i64(self, _: i64) -> Result<Self::Output> {
         Err(DataError::UnexpectedAttribute(self.expecting(), "i64".into()).into())
     }
 
-    /// Visit `f32` value.
-    fn visit_f32(self, _: f32) -> Result<Self::Output> {
+    /// Loads `f32` value.
+    fn load_f32(self, _: f32) -> Result<Self::Output> {
         Err(DataError::UnexpectedAttribute(self.expecting(), "f32".into()).into())
     }
 
-    /// Visit `f64` value.
-    fn visit_f64(self, _: f64) -> Result<Self::Output> {
+    /// Loads `f64` value.
+    fn load_f64(self, _: f64) -> Result<Self::Output> {
         Err(DataError::UnexpectedAttribute(self.expecting(), "f64".into()).into())
     }
 
-    /// Visit boolean array.
-    fn visit_seq_bool(
+    /// Loads boolean array.
+    fn load_seq_bool(
         self,
         _: impl Iterator<Item = Result<bool>>,
         _len: usize,
@@ -61,8 +51,8 @@ pub trait VisitAttribute: Sized + fmt::Debug {
         Err(DataError::UnexpectedAttribute(self.expecting(), "boolean array".into()).into())
     }
 
-    /// Visit `i32` array.
-    fn visit_seq_i32(
+    /// Loads `i32` array.
+    fn load_seq_i32(
         self,
         _: impl Iterator<Item = Result<i32>>,
         _len: usize,
@@ -70,8 +60,8 @@ pub trait VisitAttribute: Sized + fmt::Debug {
         Err(DataError::UnexpectedAttribute(self.expecting(), "i32 array".into()).into())
     }
 
-    /// Visit `i64` array.
-    fn visit_seq_i64(
+    /// Loads `i64` array.
+    fn load_seq_i64(
         self,
         _: impl Iterator<Item = Result<i64>>,
         _len: usize,
@@ -79,8 +69,8 @@ pub trait VisitAttribute: Sized + fmt::Debug {
         Err(DataError::UnexpectedAttribute(self.expecting(), "i64 array".into()).into())
     }
 
-    /// Visit `f32` array.
-    fn visit_seq_f32(
+    /// Loads `f32` array.
+    fn load_seq_f32(
         self,
         _: impl Iterator<Item = Result<f32>>,
         _len: usize,
@@ -88,8 +78,8 @@ pub trait VisitAttribute: Sized + fmt::Debug {
         Err(DataError::UnexpectedAttribute(self.expecting(), "f32 array".into()).into())
     }
 
-    /// Visit `f64` array.
-    fn visit_seq_f64(
+    /// Loads `f64` array.
+    fn load_seq_f64(
         self,
         _: impl Iterator<Item = Result<f64>>,
         _len: usize,
@@ -97,31 +87,31 @@ pub trait VisitAttribute: Sized + fmt::Debug {
         Err(DataError::UnexpectedAttribute(self.expecting(), "f64 array".into()).into())
     }
 
-    /// Visit binary value.
+    /// Loads binary value.
     ///
     /// This method should return error when the given reader returned error.
-    fn visit_binary(self, _: impl io::Read, _len: u64) -> Result<Self::Output> {
+    fn load_binary(self, _: impl io::Read, _len: u64) -> Result<Self::Output> {
         Err(DataError::UnexpectedAttribute(self.expecting(), "binary data".into()).into())
     }
 
-    /// Visit binary value on buffered reader.
+    /// Loads binary value on buffered reader.
     ///
     /// This method should return error when the given reader returned error.
-    fn visit_binary_buffered(self, reader: impl io::BufRead, len: u64) -> Result<Self::Output> {
-        self.visit_binary(reader, len)
+    fn load_binary_buffered(self, reader: impl io::BufRead, len: u64) -> Result<Self::Output> {
+        self.load_binary(reader, len)
     }
 
-    /// Visit string value.
+    /// Loads string value.
     ///
     /// This method should return error when the given reader returned error.
-    fn visit_string(self, _: impl io::Read, _len: u64) -> Result<Self::Output> {
+    fn load_string(self, _: impl io::Read, _len: u64) -> Result<Self::Output> {
         Err(DataError::UnexpectedAttribute(self.expecting(), "string data".into()).into())
     }
 
-    /// Visit string value on buffered reader.
+    /// Loads string value on buffered reader.
     ///
     /// This method should return error when the given reader returned error.
-    fn visit_string_buffered(self, reader: impl io::BufRead, len: u64) -> Result<Self::Output> {
-        self.visit_string(reader, len)
+    fn load_string_buffered(self, reader: impl io::BufRead, len: u64) -> Result<Self::Output> {
+        self.load_string(reader, len)
     }
 }